refactor(v1): rename clients.openai module -> clients.proxy
The module's client is now ProxyClient (forwards 1:1), so name the module for what
it is. Pure rename + import updates (config, clients __init__, renderer, the
interception server, the default harness); also tidies the module docstring and
shortens the get_response docstring.
