+++ verifiers/v1/clients/__init__.py
@@ -8,7 +8,7 @@ from verifiers.v1.clients.config import (
     RendererClientConfig,
     resolve_client,
 )
-from verifiers.v1.clients.openai import ProxyClient
+from verifiers.v1.clients.proxy import ProxyClient
 from verifiers.v1.clients.renderer import RendererClient
 
 __all__ = [
