+++ verifiers/v1/clients/config.py
@@ -17,7 +17,7 @@ from pydantic_config import BaseConfig
 from renderers import RendererConfig
 
 from verifiers.v1.clients.client import Client
-from verifiers.v1.clients.openai import ProxyClient
+from verifiers.v1.clients.proxy import ProxyClient
 from verifiers.v1.clients.renderer import RendererClient
 
 PRIME_INFERENCE_HOST = "pinference.ai"
